[ur-4mfwg] Replace ticket create overlay flow with $EDITOR + frontmatter approach
Shift+C now breaks out of the TEA loop, tears down the terminal, spawns
$EDITOR with a frontmatter template (title/priority/body), parses the
result via v1's parse_ticket_file(), re-initializes the terminal, and
shows the CreateActionMenu overlay for the final action. Project is
auto-derived from parent ticket project or project_filter. Removes the
multi-step ProjectInput/TitleInput overlay create flow and the
CreateTicketState from the model. Adds body field to PendingTicket so
editor content flows through to ticket creation.

+++ crates/urui/src/v2/pages/ticket_detail.rs
@@ -317,7 +317,7 @@ fn handle_detail_priority(model: Model) -> (Model, Vec<Cmd>) {
 /// Close the selected child, prompting for force close if it has open children.
 fn handle_detail_close(model: Model) -> (Model, Vec<Cmd>) {
     if let Some(ticket) = selected_child(&model) {
-        let open_children = (ticket.children_total - ticket.children_completed) as i32;
+        let open_children = ticket.children_total - ticket.children_completed;
         if open_children > 0 {
             let msg = Msg::Overlay(OverlayMsg::OpenForceCloseConfirm {
                 ticket_id: ticket.id.clone(),
